requirements.txt
numpy
scipy
IPython
Pillow
matplotlib
scikit-learn